change shop detail screen
ナビゲーション機能でタブ切り替えを行なっていた詳細ページを、NativeBaseのテンプレートに変更。
この変更にに伴い、マップ専用スクリーンを削除し、詳細スクリーンに統一する。
