+++ src/navigator.vue
@@ -9,7 +9,6 @@ import HomeScreen from "./screens/HomeScreen.vue"
 import SigninScreen from "./screens/SigninScreen.vue"
 import SignupScreen from "./screens/SignupScreen.vue"
 import DetailScreen from "./screens/DetailScreen.vue"
-import MapScreen from "./screens/MapScreen.vue"
 import SearchScreen from "./screens/SearchScreen.vue"
 import MypageScreen from "./screens/MypageScreen.vue"
 import { Root } from "native-base"
@@ -20,19 +19,6 @@ import {
   createMaterialTopTabNavigator
 } from "vue-native-router"
 
-const DetailTab = createMaterialTopTabNavigator(
-  {
-    Detail: DetailScreen,
-    Map: MapScreen
-  },
-  { 
-    initialRouteName: 'Detail',
-    tabBarOptions: {
-      style: { backgroundColor: '#FFCC33' }
-    }
-  }
-)
-
 const StackNavigator = createStackNavigator(
   {
     Home: HomeScreen,
@@ -40,7 +26,7 @@ const StackNavigator = createStackNavigator(
     Signup: SignupScreen,
     Search: SearchScreen,
     Mypage: MypageScreen,
-    DetailTab
+    Detail: DetailScreen
   },
   {
     initialRouteName: 'Home',
